Cleanup and added vive
Added some comments to compare_datasets and removed some unnecessary lines.
Added functionality to use the vive as reference system (yields worse results than krypton so far).
LogVive contains the files to get the data from steamVR to csv files and to store the pose matrices in a numpy object.